requirements.txt
langchain==0.0.350
langchain-openai==0.0.2
langgraph==0.0.35
tavily-python==0.3.9
openai==1.13.3
python-dotenv==1.0.1
pydantic==2.6.3
azure-ai-inference==1.0.0b2
azure-core>=1.30.0
markdown==3.5.1
